backend/mySearchEngine/requirements.txt
django>=3.1.4
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0
requests>=2.25.0 